Fetch method now in redux, added scripts
