+++ js/components/utils/CoinAdapter.js
@@ -2,7 +2,6 @@
  * @providesModule CoinAdapter
  * @flow
  */
-import { convertJsonTypes } from 'CoinUtil';
 import { getStore } from 'GlobalStore'; 
 
 export function adaptCoinData(data) {
@@ -38,4 +37,34 @@ export function getSymbol() {
         default:
             return '£';
     }
+}
+
+export function convertJsonTypes(item) {
+    const currencyType = getStore().getState().coinReducer.currencyType;
+
+    let adaptedCoin = {};
+
+    switch(currencyType) {
+        case('gbp'):
+            adaptedCoin.price = item.quotes.GBP.price;
+            adaptedCoin.timePeriod = item.quotes.GBP.volume_24h;
+            adaptedCoin.percentageChange = item.quotes.GBP.percent_change_24h;
+            break;
+        case('usd'):
+            adaptedCoin.price = item.quotes.USD.price;
+            adaptedCoin.timePeriod = item.quotes.USD.volume_24h;
+            adaptedCoin.percentageChange = item.quotes.USD.percent_change_24h;
+            break;
+        case('eur'):
+            adaptedCoin.price = item.quotes.EUR.price;
+            adaptedCoin.timePeriod = item.quotes.EUR.volume_24h;
+            adaptedCoin.percentageChange = item.quotes.EUR.percent_change_24h;
+            break;
+        default:
+            adaptedCoin.price = item.quotes.GBP.price;
+            adaptedCoin.timePeriod = item.quotes.GBP.volume_24h;
+            adaptedCoin.percentageChange = item.quotes.GBP.percent_change_24h;
+    }
+
+    return adaptedCoin;
 }
