+++ js/redux/CoinActionTypes.js
@@ -5,3 +5,4 @@
 
 export const CHANGE_CURRENCY_TYPE = 'CURRENCY_TYPE';
 export const CHANGE_NUMBER_COINS = 'NUMBER_COINS';
+export const GET_COIN_DATA = 'GET_COIN_DATA';
