+++ package.json
@@ -3,8 +3,9 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "test": "jest",
+    "device": "adb reverse tcp:8081 tcp:8081",
+    "start": "react-native start -- --clear-cache"
   },
   "dependencies": {
     "react": "^16.0.0-alpha.12",
